Add Codespaces configuration
With this change one can use githubs codespaces feature that will create a development environment on a vm. A click in the Github UI will create an instance. Its free for all(single instance) except cooperations.

With a simple click on now ends up with a fully functional development environment. This includes:

- UAA Instance
- Minio S3 Blobstore
- Nginx Reverse Proxy for package uploads
- Mariadb and Postgresdb
- A service broker with configureable behaviour (CATS Broker)
- Run Configurations for VSCode and Intellij/Rubymine to start right away
- Tunneling SSH from the container to localhost to be able to connect with Intellij/Rubymines new remote ssh feature

At startup the DBs are set up, migrated and seeded and also all bundle packages get installed. The setup of a codespaces takes rughly 5 minutes. Once a codespace is stopped it wont get recreated at startup so startup times are nearly instant. Also changes to the environment are persisted between restarts. Only if one choses to throw away a codespace and recreate it, thats when the build times come on top of starting times and any temporary change to the environment is gone.

We will use this as our main development environment for CC. In case it gets unmaintained its a simple delete of a folder that is just used by the github codespaces functionality.

Also this enables super simple start on developing in CC for most endpoints (excluding diego and loggregator functionality for now).